Modify bug of Map method when source type is ValueType.

+++ HigLabo.Test/HigLabo.Mapper.Test/Core/Vector2.cs
@@ -8,10 +8,10 @@ namespace HigLabo.Mapper.Test
 {
     public struct Vector2
     {
-        public Decimal X { get; private set; }
-        public Decimal Y { get; private set; }
+        public Int32 X { get; private set; }
+        public Int32 Y { get; private set; }
 
-        public Vector2(Decimal x, Decimal y)
+        public Vector2(Int32 x, Int32 y)
         {
             this.X = x;
             this.Y = y;
